+++ packages/provider-mongo/test/example-types/exists.js
@@ -2,7 +2,21 @@ let { expect } = require('chai')
 let _ = require('lodash/fp')
 let exists = require('../../src/example-types/exists')
 
+let node = {
+  type: 'exists',
+  field: 'test',
+}
+
 describe('exists', () => {
+  describe('exists.hasValue', () => {
+    it('Should detect a boolean value only', () => {
+      expect(exists.hasValue({ ...node, value: true })).to.be.true
+      expect(exists.hasValue({ ...node, value: false })).to.be.true
+      expect(exists.hasValue(node)).to.be.false
+      expect(exists.hasValue({ ...node, value: null })).to.be.false
+      expect(exists.hasValue({ ...node, value: undefined })).to.be.false
+    })
+  })
   describe('exists.filter', () => {
     it('If a value is provided, use $and', () => {
       expect(